XML file found in downloads section of original Google Code Project.
Author - Sedna. If the author is reading this, I have transferred your project to GitHub as Google Code was shutting down, and I wanted to preserve your work. I was not sure whether you were active or not. No copyright infringement is or was intended, and please contact me if do not want your project hosted on my account, I shall promptly delete the repository.